pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

